Set up JUnit the way it is officially documented
I tried hacking on the project and found the tests not to be runnable in
Eclipse while they do work when run using `./gradlew test`.

I have encountered that before, checked the official docs both on the
JUnit website and the Gradle user guide and found the setup not to be
done as recommended. With this update, it keeps working on the command
line but also works in Eclipse. Running the tests in IntelliJ/Android
Studio also still works.

See https://junit.org/junit5/docs/current/user-guide/#running-tests-build-gradle-engines-configure
The paragraph just before that headline also mentiones that this is
required for IDEs.

Also https://docs.gradle.org/current/userguide/java_testing.html#sec:java_testing_basics
has the same kind of setup as an example.